Add test for Helper::logger
Unfortunately, this test case covers only that the helper does not
log when log mode is disabled.

This is due to the fact that it's
seemingly not possible to set different function call expectations
between function calls in BrainMonkey, as Pest has issues with running
tests in isolation. I don't see a straightforward way to mock error_log
to allow this to be tested properly.