Add mimetype validation helpers to AbstractValidation
This commit also extracts parsing the allowed filetypes string
and getting file extension from filename to new private methods,
as they're shared logic.

It also provides a filter to allow developers to force mimetype
validation to be based upon filesystem values, instead of
falling back to POST-provided values in that case. This setting
isn't (and I believe shouldn't be) configurable in wp-admin.